sage_maker_fastapi/requirements.txt
nltk
fastapi
uvicorn[standard]
boto3
python-dotenv
gunicorn